requirement.txt
torch==2.3.1+cu118  
torchvision==0.18.1+cu118  
torchaudio==2.3.1  
ultralytics>=8.2.0  
opencv-python>=4.5.0  
numpy<=2.0.0  
pyyaml>=5.4.0  
albumentations==0.4.3  
hydra-core>=1.0.0  
pytorch-lightning>=1.0.0 
kornia>=0.5.0
webdataset  
scikit-image 
scikit-learn 
easydict  
omegaconf>=2.0.0  